Implementació de lectura de BD en modul de partides, falta CRUD.

+++ src/main/java/com/example/cleverp/model/Partides.java
@@ -12,6 +12,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+import jakarta.validation.constraints.Null;
 import java.sql.Date;
 import java.util.List;
 import lombok.Data;
@@ -35,11 +36,11 @@ public class Partides {
     @Column(name = "aforament")
     private Integer aforament;
     @Column(name = "EquipA_id")
-    private long EquipA_id;
+    private Integer EquipA_id;
     @Column(name = "EquipB_id")
-    private long EquipB_id;
+    private Integer EquipB_id;
     @Column(name = "guanyador_sorteig")
-    private long guanyador_sorteig;
+    private Integer guanyador_sorteig;
     @Column(name = "producte_sorteig")
     private String producte_sorteig;
     
@@ -49,5 +50,4 @@ public class Partides {
     @OneToMany
     @JoinColumn(name = "EquipB_id")
     private List<Equip> EquipsB;
-    
 }
